refactor : Instagram dto 구조 변경
