+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/Instagram/dto/InstagramRequest.java
@@ -1,4 +1,4 @@
-package com.techeerpicture.dto;
+package com.techeerpicture.TecheerPicture.Instagram.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
